# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/sale_order_item": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of sale order items or a specific sale order item by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale_order_items"
                ],
                "summary": "Get sale order items",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sale Order Item ID",
                        "name": "sale_order_itemID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.SaleOrderItem"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new sale order item with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale_order_items"
                ],
                "summary": "Create a new sale order item",
                "parameters": [
                    {
                        "description": "Sale Order Item",
                        "name": "sale_order_item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SaleOrderItem"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.SaleOrderItem"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sale_order_item/sale_order": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of sale order items by sale order ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale_order_items"
                ],
                "summary": "Get sale order items by sale order ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sale Order ID",
                        "name": "sale_order_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.SaleOrderItem"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sale_order_item/{sale_order_itemID}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update an sale order item with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale_order_items"
                ],
                "summary": "Update an sale order item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sale Order Item ID",
                        "name": "sale_order_itemID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sale Order Item",
                        "name": "sale_order_item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SaleOrderItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SaleOrderItem"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete an sale order item by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sale_order_items"
                ],
                "summary": "Delete an sale order item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sale Order Item ID",
                        "name": "sale_order_itemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sale Order Item deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Sale Order Item not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.SaleOrderItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "sale_order_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}